Easier Dll loads and minor bug fix:
The LibraryManager now searches for the init methods of all public classes in the class library fitting the name described in the interface, so multiple libraries may be loaded from the same dll and there is no depencence on the order of your class files.

A minor bug where an invalid command could leave the parser was also fixed.