Comments for moving forward and some small tweaks
Each piece of the CommandProcessing folder has been commented for moving forward with the project.
Additional small tweaks:
 - To keep modifications to the window moderated by CommandExecutor, the BuiltinHandler, BuiltinLibrary,
    ICommandHandler, and ICommandLibrary code has been changed to only access the CommandExecutor
    and the LibraryManager.
-  Continuing with this idea, the necessary reference to the CommandProcessor held by the LibraryManager
    has been changed from a Property to a private Field. This is because a reference to the LibraryManager is held
    by its ICommandLibrary objects.

With these changes and the additional comments, the code is now ready to be ported to .Net Core. After this,
some work (noted in the comments) is still left to be done to make this version completely stable.